09_Apigee/apiServer/requirements.txt
Flask
requests
firebase_admin
google-cloud-pubsub
google-auth